Fix testlibrary_tests/TestMutuallyExclusivePlatformPredicates.java
